+++ test/features/remote.js
@@ -14,7 +14,7 @@ define('features/remote', [
         // Restore ParsleyExtend from remote
         window.ParsleyExtend = window._remoteParsleyExtend;
         window.ParsleyConfig = window._remoteParsleyConfig;
-        window.ParsleyValidator.init(window.ParsleyConfig.validators);
+        window.ParsleyValidator.init(window.ParsleyConfig.validators, window.ParsleyConfig.i18n);
       });
       it('should have window.ParsleyExtend defined', function () {
         expect(window.ParsleyExtend).not.to.be(undefined);